+++ demos/browser/app/meeting/meeting.ts
@@ -22,6 +22,7 @@ import {
   MeetingSessionStatus,
   MeetingSessionStatusCode,
   MeetingSessionVideoAvailability,
+  POSTRequestLogger,
   ScreenMessageDetail,
   ScreenShareFacadeObserver,
   TimeoutScheduler,
@@ -128,7 +129,7 @@ export class DemoMeetingApp implements AudioVideoObserver, DeviceChangeObserver 
 
   // feature flags
   enableWebAudio = false;
-
+  logger: any = null;
   constructor() {
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
     (global as any).app = this;
@@ -442,22 +443,26 @@ export class DemoMeetingApp implements AudioVideoObserver, DeviceChangeObserver 
       }
       new AsyncScheduler().start(async () => {
         (buttonMeetingEnd as HTMLButtonElement).disabled = true;
+        (buttonMeetingLeave as HTMLButtonElement).disabled = true;
         await this.endMeeting();
         this.leave();
-        (buttonMeetingEnd as HTMLButtonElement).disabled = false;
         // @ts-ignore
         window.location = window.location.pathname;
+        (buttonMeetingEnd as HTMLButtonElement).disabled = false;
+        (buttonMeetingLeave as HTMLButtonElement).disabled = false;
       });
     });
 
     const buttonMeetingLeave = document.getElementById('button-meeting-leave');
     buttonMeetingLeave.addEventListener('click', _e => {
       new AsyncScheduler().start(async () => {
+        (buttonMeetingEnd as HTMLButtonElement).disabled = true;
         (buttonMeetingLeave as HTMLButtonElement).disabled = true;
         this.leave();
-        (buttonMeetingLeave as HTMLButtonElement).disabled = false;
         // @ts-ignore
         window.location = window.location.pathname;
+        (buttonMeetingEnd as HTMLButtonElement).disabled = false;
+        (buttonMeetingLeave as HTMLButtonElement).disabled = false;
       });
     });
   }
@@ -548,11 +553,16 @@ export class DemoMeetingApp implements AudioVideoObserver, DeviceChangeObserver 
     }
   }
 
+
   async initializeMeetingSession(configuration: MeetingSessionConfiguration): Promise<void> {
-    const logger = new ConsoleLogger('SDK', LogLevel.DEBUG);
-    const deviceController = new DefaultDeviceController(logger);
+    if (!(DemoMeetingApp.BASE_URL.includes('localhost') || DemoMeetingApp.BASE_URL.includes("127.0.0.1"))) {
+      this.logger = new POSTRequestLogger('SDK', configuration, 75, 1750, `${DemoMeetingApp.BASE_URL}logs`, LogLevel.INFO);
+    } else {
+      this.logger = new ConsoleLogger('SDK', LogLevel.DEBUG);
+    }
+    const deviceController = new DefaultDeviceController(this.logger);
     configuration.enableWebAudio = this.enableWebAudio;
-    this.meetingSession = new DefaultMeetingSession(configuration, logger, deviceController);
+    this.meetingSession = new DefaultMeetingSession(configuration, this.logger, deviceController);
     this.audioVideo = this.meetingSession.audioVideo;
 
     this.audioVideo.addDeviceChangeObserver(this);